Refactoring to reduce file lengths. Small changes.
